+++ srml/aura/src/lib.rs
@@ -54,9 +54,7 @@ use rstd::{result, prelude::*};
 use parity_codec::Encode;
 use srml_support::{decl_storage, decl_module, Parameter, storage::StorageValue, traits::Get};
 use primitives::{
-	traits::{
-		Convert, SaturatedConversion, Saturating, Zero, One, Member, IsMember, TypedKey,
-	},
+	traits::{SaturatedConversion, Saturating, Zero, One, Member, IsMember, TypedKey},
 	generic::DigestItem,
 };
 use timestamp::OnTimestampSet;
