+++ srml/babe/src/lib.rs
@@ -25,7 +25,10 @@ use rstd::{result, prelude::*};
 use srml_support::{decl_storage, decl_module, StorageValue, traits::FindAuthor, traits::Get};
 use timestamp::{OnTimestampSet};
 use primitives::{generic::DigestItem, ConsensusEngineId};
-use primitives::traits::{IsMember, SaturatedConversion, Saturating, RandomnessBeacon, Convert, TypedKey};
+use primitives::traits::{
+	CurrentSessionKeys, IsMember, SaturatedConversion, Saturating, RandomnessBeacon, Convert,
+	TypedKey,
+};
 #[cfg(feature = "std")]
 use timestamp::TimestampInherentData;
 use parity_codec::{Encode, Decode};
@@ -107,11 +110,11 @@ impl ProvideInherentData for InherentDataProvider {
 	}
 }
 
-pub trait Trait: timestamp::Trait + session::Trait {
+pub trait Trait: timestamp::Trait {
 	type EpochDuration: Get<u64>;
 
-	/// Interface for interacting with a session module.
-	type SessionInterface: self::SessionInterface<Self::AccountId>;
+	/// Retrieve the current session keys.
+	type CurrentSessionKeys: CurrentSessionKeys<Self::AccountId>;
 }
 
 /// The length of the BABE randomness
@@ -214,30 +217,6 @@ impl<T: Trait> IsMember<AuthorityId> for Module<T> {
 	}
 }
 
-/// Means for interacting with a specialized version of the `session` trait.
-///
-/// This is needed because `Staking` sets the `ValidatorIdOf` of the `session::Trait`
-pub trait SessionInterface<AccountId>: system::Trait {
-	/// Get the keys of the current session.
-	fn current_keys<Key: Decode + Default + TypedKey>() -> Vec<(AccountId, Key)>;
-}
-
-impl<T: Trait> SessionInterface<<T as system::Trait>::AccountId> for T where
-	T::AccountIdOf: Convert<<T as session::Trait>::ValidatorId, Option<<T as system::Trait>::AccountId>>,
-{
-	fn current_keys<Key>() -> Vec<(<T as system::Trait>::AccountId, Key)>
-		where Key: Decode + Default + TypedKey
-	{
-		<session::Module<T>>::get_current_keys::<Key>()
-			.into_iter()
-			.map(|(validator_id, keys)| {
-				let account_id = T::AccountIdOf::convert(validator_id).unwrap_or_default();
-				(account_id, keys)
-			})
-			.collect()
-	}
-}
-
 /// A `Convert` implementation that finds the babe authority id of the given controller
 /// account, if any.
 pub struct AuthorityIdOf<T, Id>(rstd::marker::PhantomData<T>, rstd::marker::PhantomData<Id>);
@@ -248,7 +227,7 @@ impl<T, AuthorityId> Convert<T::AccountId, Option<AuthorityId>> for AuthorityIdO
 		AuthorityId: Decode + Default + TypedKey,
 {
 	fn convert(account_id: T::AccountId) -> Option<AuthorityId> {
-		let keys = T::SessionInterface::current_keys::<AuthorityId>();
+		let keys = T::CurrentSessionKeys::current_keys::<AuthorityId>();
 		let maybe_authority_id = keys
 			.into_iter()
 			.find(|(id, _)| {
